+++ tests/resources/views/welcome.blade.php
@@ -1,0 +1,7 @@
+<x-mail::message>
+The brand is {{ brand('name') }}.
+
+<x-mail::button :url="'https://example.test'">
+Get started
+</x-mail::button>
+</x-mail::message>